refactor: consolidate FieldChange, rune-safe truncation, audit From/To, split edit.go
1. Eliminate duplicate FieldChange struct — use audit.FieldChange as the
   single source of truth. UpdateDirect now returns []audit.FieldChange
   directly, removing the manual conversion loop in the command layer.

2. truncateForAudit now truncates on rune boundaries (not byte boundaries)
   to avoid splitting multi-byte UTF-8 characters.

3. Audit edit events now populate From/To fields when status changes,
   matching the query pattern used by move events.

4. Extract picker/UI helpers (pickIntent, resolveIntentByPartialID,
   formatIntentPreview, truncateID) into edit_picker.go. edit.go drops
   from 473 to 355 lines.
